enable outputs

+++ src/IOManager.h
@@ -1,8 +1,7 @@
-// #include <CiiiXo.h>
+#include <CiiiXo.h>
 
 /******************DEFINE*******************/
 #define PIN_RESET 37
-
 
 /******************PCA9698*******************/
 #define PIN_INT_PCA9698 35
@@ -17,15 +16,16 @@ const uint8_t MODULE_SIZE = 40;
 const uint8_t MAX_INPUT_MODULES = 2;
 const uint8_t MAX_OUTPUT_MODULES = 4;
 
-enum IOType {INPUT_MODULE, OUTPUT_MODULE};
+enum IOType { INPUT_MODULE,
+              OUTPUT_MODULE };
 
 void configureIO();
 void beginIO();
-void setPin(int pin, int value);
+void setPin(uint8_t pin, uint8_t value);
 uint8_t readPin(int pin);
-uint8_t readPin(uint8_t pin, IOType io_type); // Ability to read output pins
+uint8_t readPin(uint8_t pin, IOType io_type);  // Ability to read output pins
 void readAll(uint8_t data[MODULE_SIZE * INPUT_MODULE]);
-void readAll(uint8_t data[MODULE_SIZE * INPUT_MODULE], IOType io_type); // Ability to read output pins
+void readAll(uint8_t data[MODULE_SIZE * INPUT_MODULE], IOType io_type);  // Ability to read output pins
 bool ioChangeDetected();
 void ioChanges(uint8_t data[MODULE_SIZE * INPUT_MODULE]);
 bool detectDevice(uint8_t address);
